+++ doc/releases.rst
@@ -58,6 +58,47 @@ requires Java 8.
           |   0.100 |   0.450 |   0.193 |   0.200 |  12.456 |   0.200 |   0.000 |
           -----------------------------------------------------------------------
 
+    SLDWriter has optional named parameters for the number of spaces to use while indenting and
+    whether to export default values or not::
+
+        Symbolizer sym = new Fill("wheat") + new Stroke("brown")
+        SLDWriter writer = new SLDWriter()
+        String sld = writer.write(sym, exportDefaultValues: true, indentation: 4)
+        println sld
+
+        <?xml version="1.0" encoding="UTF-8"?>
+        <sld:StyledLayerDescriptor xmlns="http://www.opengis.net/sld" xmlns:sld="http://www.opengis.net/sld" xmlns:ogc="http://www.opengis.net/ogc" xmlns:gml="http://www.opengis.net/gml" version="1.0.0">
+            <sld:UserLayer>
+                <sld:LayerFeatureConstraints>
+                    <sld:FeatureTypeConstraint/>
+                </sld:LayerFeatureConstraints>
+                <sld:UserStyle>
+                    <sld:Name>Default Styler</sld:Name>
+                    <sld:FeatureTypeStyle>
+                        <sld:Name>name</sld:Name>
+                        <sld:Rule>
+                            <sld:PolygonSymbolizer>
+                                <sld:Fill>
+                                    <sld:CssParameter name="fill">#f5deb3</sld:CssParameter>
+                                    <sld:CssParameter name="fill-opacity">1.0</sld:CssParameter>
+                                </sld:Fill>
+                            </sld:PolygonSymbolizer>
+                            <sld:LineSymbolizer>
+                                <sld:Stroke>
+                                    <sld:CssParameter name="stroke">#a52a2a</sld:CssParameter>
+                                    <sld:CssParameter name="stroke-linecap">butt</sld:CssParameter>
+                                    <sld:CssParameter name="stroke-linejoin">miter</sld:CssParameter>
+                                    <sld:CssParameter name="stroke-opacity">1.0</sld:CssParameter>
+                                    <sld:CssParameter name="stroke-width">1</sld:CssParameter>
+                                    <sld:CssParameter name="stroke-dashoffset">0.0</sld:CssParameter>
+                                </sld:Stroke>
+                            </sld:LineSymbolizer>
+                        </sld:Rule>
+                    </sld:FeatureTypeStyle>
+                </sld:UserStyle>
+            </sld:UserLayer>
+        </sld:StyledLayerDescriptor>
+
 1.8.0
 -----
 The 1.8.0 release of GeoScript is built on Grooovy 2.4.7, GeoTools 16.0, and the Java Topology Suite 1.13 and